requirements.txt
torch>=2.0.0
transformers>=4.40.0
pyyaml>=5.0
pillow
cairosvg
fire
tqdm
numpy>=1.26
piqa
vllm>=0.4.0
